Move gate approximations to gate package
